refactor: move mantine providers to theme provider
